found and fixed a bug with the register/login functions
Found bug: newly created users were only saved to database, but not loaded to memory, so that they could not log in until after server restart. This was a result of the DB only being read to memory at server start. added on top of this the problem, that every time the server was restart, the users created during this session would work, but the users in the original session would be deleted/overwritten by new users.

Bug Fix:
Server still reads the data from DB only at server start. however, when a new user is created he (the user data) is only added to the "users" variable, and then the whole "users" variable is the saved to the database asynchronously so that the server is not interrupted.
I could probably have created a function that only saves user data on server shutdown, however this could create a problem if a sudden server crash occurs. By using my method the data is stored after every change.